+++ pkg/namespace/namespaces.go
@@ -1,22 +1,24 @@
 package namespace
 
 import (
-	log "github.com/sirupsen/logrus"
+	"github.com/iver-wharf/wharf-core/pkg/logger"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 )
+
+var log = logger.New()
 
 type Namespaces struct {
 	Kubeconfig *rest.Config
 }
 
 func (n Namespaces) SetupNamespaces(namespaces []string) error {
-	log.Traceln("SetupNamespace called")
+	log.Debug().Message("SetupNamespace called")
 
 	clientSet, err := kubernetes.NewForConfig(n.Kubeconfig)
 	if err != nil {
-		log.Fatal(err)
+		log.Panic().WithError(err).Message("Failed loading kube-config.")
 	}
 
 	nsClient := clientSet.CoreV1().Namespaces()
@@ -24,18 +26,20 @@ func (n Namespaces) SetupNamespaces(namespaces []string) error {
 		ns, err := nsClient.Get(name, metav1.GetOptions{})
 
 		if ns.Name == "" {
-			log.WithField("name", name).Infoln("Creating namespace")
+			log.Debug().WithString("namespace", name).Message("Creating namespace.")
 
 			ns.Name = name
 			_, err = nsClient.Create(ns)
 			if err != nil {
-				log.WithError(err).Errorln("Error creating namespace")
+				log.Error().WithError(err).Message("Failed to create namespace.")
+			} else {
+				log.Info().WithString("namespace", name).Message("Created namespace.")
 			}
 
 			continue
 		}
 
-		log.WithField("name", name).Traceln("Namespace found!")
+		log.Info().WithString("namespace", name).Message("Namespace already exists.")
 	}
 
 	return nil