+++ pkg/run/run.go
@@ -5,12 +5,14 @@ import (
 	"sync"
 
 	"github.com/iver-wharf/wharf-api-client-go/pkg/wharfapi"
-	log "github.com/sirupsen/logrus"
 	"github.com/iver-wharf/wharf-cmd/pkg/core/buildclient"
 	"github.com/iver-wharf/wharf-cmd/pkg/core/kubernetes"
 	"github.com/iver-wharf/wharf-cmd/pkg/core/wharfyml"
+	"github.com/iver-wharf/wharf-core/pkg/logger"
 	"k8s.io/client-go/rest"
 )
+
+var log = logger.New()
 
 type Runner struct {
 	Kubeconfig  *rest.Config
@@ -27,16 +29,21 @@ func NewRunner(kubeconfig *rest.Config, authHeader string) Runner {
 }
 
 func (r Runner) Run(path, environment, namespace, stageName string, buildID int, builtinVars map[wharfyml.BuiltinVar]string) error {
-	log.WithFields(log.Fields{
-		"path":      path,
-		"namespace": namespace,
-		"stage":     stageName,
-		"buildID":   buildID,
-	}).Traceln("Run called")
+	var withRunMeta = func(ev logger.Event) logger.Event {
+		return ev.
+			WithString("path", path).
+			WithString("namespace", namespace).
+			WithString("stage", stageName).
+			WithInt("buildID", buildID)
+	}
+	withRunMeta(log.Debug()).
+		Message("Run called.")
 
 	def, err := wharfyml.Parse(path, builtinVars)
 	if err != nil {
-		log.WithError(err).Errorln("Failed to parse wharf-ci file!")
+		withRunMeta(log.Error()).
+			WithError(err).
+			Message("Failed to parse wharf-ci file.")
 		return fmt.Errorf("run: parse definition: %w", err)
 	}
 
@@ -48,22 +55,27 @@ func (r Runner) RunDefinition(
 	environment, namespace, stageName string,
 	buildID int,
 	builtinVars map[wharfyml.BuiltinVar]string) error {
+	var withRunMeta = func(ev logger.Event) logger.Event {
+		return ev.
+			WithString("namespace", namespace).
+			WithString("stage", stageName).
+			WithInt("buildID", buildID)
+	}
 	stage, err := definition.GetStageWithReplacement(stageName, environment)
 	if err != nil {
-		log.WithError(err).
-			WithFields(log.Fields{"stage": stageName, "environment": environment}).
-			Errorln("Stage not found")
+		withRunMeta(log.Error()).
+			WithError(err).
+			Message("Stage not found.")
 		return fmt.Errorf("stage %q not found in definition", stageName)
 	}
 
-	err = r.buildClient.PostLogWithStatus(uint(buildID), "run definition", wharfapi.BuildScheduling)
-	if err != nil {
-		log.WithError(err).Errorln("Unable to update build status")
-	}
+	r.postLogWithStatus(buildID, "Parsed run definition.", wharfapi.BuildScheduling)
 
 	podClient, err := kubernetes.NewPodClient(namespace, r.Kubeconfig)
 	if err != nil {
-		log.WithError(err).Fatalln("Error getting new pod client")
+		withRunMeta(log.Error()).
+			WithError(err).
+			Message("Error getting new pod client")
 		return fmt.Errorf("run definition: get new pod client: %w", err)
 	}
 
@@ -71,19 +83,15 @@ func (r Runner) RunDefinition(
 	for _, step := range stage.Steps {
 		pod, err := podClient.CreatePod(step, stage, buildID)
 		if err != nil {
-			log.WithError(err).Fatalln("Failed to create pod")
+			withRunMeta(log.Error()).
+				WithError(err).
+				Message("Failed to create pod.")
 
-			err = r.buildClient.PostLogWithStatus(uint(buildID), "Unable to create pod.", wharfapi.BuildFailed)
-			if err != nil {
-				log.WithError(err).Errorln("Unable to update build status")
-			}
+			r.postLogWithStatus(buildID, "Unable to create pod", wharfapi.BuildFailed)
 			return fmt.Errorf("run definition: create pod: %w", err)
 		}
 
-		err = r.buildClient.PostLogWithStatus(uint(buildID), fmt.Sprintf("Pod %s created.", pod.Name), wharfapi.BuildRunning)
-		if err != nil {
-			log.WithError(err).Errorln("Unable to update build status")
-		}
+		r.postLogWithStatus(buildID, fmt.Sprintf("Pod %s created.", pod.Name), wharfapi.BuildRunning)
 
 		wg.Add(1)
 		go func() {
@@ -93,10 +101,7 @@ func (r Runner) RunDefinition(
 			go podClient.ReadLogsFromPod(logChannel, pod.Name)
 
 			for message := range logChannel {
-				err := r.buildClient.PostLog(uint(buildID), message)
-				if err != nil {
-					log.WithError(err).Errorln("Unable to post log")
-				}
+				r.postLog(buildID, message)
 			}
 
 			done, err := podClient.WaitUntilPodFinished(pod.Name)
@@ -106,20 +111,38 @@ func (r Runner) RunDefinition(
 
 			err = podClient.DeletePod(pod.Name)
 			if err != nil {
-				log.WithError(err).WithField("pod name", pod.Name).Errorln("Unable to delete the pod.")
+				withRunMeta(log.Error()).
+					WithError(err).
+					WithString("pod", pod.Name).
+					Message("Failed to delete pod.")
 			}
 
-			err = r.buildClient.PostLog(uint(buildID), fmt.Sprintf("Pod %s is deleted.", pod.Name))
-			if err != nil {
-				log.WithError(err).Errorln("Unable to post build log")
-			}
+			r.postLog(buildID, fmt.Sprintf("Pod %s is deleted.", pod.Name))
 		}()
 	}
 
 	wg.Wait()
-	err = r.buildClient.PostLogWithStatus(uint(buildID), "Build completed", wharfapi.BuildCompleted)
-	if err != nil {
-		log.WithError(err).Errorln("Unable to update build status")
-	}
+	r.postLogWithStatus(buildID, "Build completed.", wharfapi.BuildCompleted)
 	return nil
+}
+
+func (r Runner) postLog(buildID int, message string) {
+	err := r.buildClient.PostLog(uint(buildID), message)
+	if err != nil {
+		log.Error().
+			WithError(err).
+			WithInt("build", buildID).
+			Message("Failed to post log.")
+	}
+}
+
+func (r Runner) postLogWithStatus(buildID int, message string, status wharfapi.BuildStatus) {
+	err := r.buildClient.PostLogWithStatus(uint(buildID), message, status)
+	if err != nil {
+		log.Error().
+			WithError(err).
+			WithInt("build", buildID).
+			WithString("newStatus", status.String()).
+			Message("Failed to update build status.")
+	}
 }